colony: tvl from colony dex (forked uniswap v2) pairs
